requirements.txt
asgiref==3.7.2
Django==3.2.11
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
